tool tips
I have an x and trying to get it close. it only closes the first
tooltip of the page for some reason. possibly because i have them
initialized as an id and not a class?

the other thing i can do is make it a hover so it will just go away
when you mouse off and revert to clicks on mobile.

we should just update bootstrap for redcap and let me redo the
tooltips. this is the old version and it has a lot less features.

#15